requirements.txt
pytest>=7.0.0
pytest-cov>=3.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
pyyaml>=6.0
gitpython>=3.1.0
joblib>=1.1.0
colorlog>=6.6.0
